Properly initialize the role acceptances model
The role acceptances model was using the syntax `id = ` in
the class itself, which when compiled to JavaScript meant
the id was being stored on the class object, not
on instances of it. Therefore, the `id` would only ever
be set once.

Also, moved the code that determined the `id` from the model to
the view, since the model should not be cognizant of nor care
about the url.