Replace TOC numbering, handle sequential chapters
CSS numbering cannot work for sequential chapters, because unexpanded
sections are not on the page and so cannot be counted. This patch
applies a “chapter” attribute to TOC model entries and includes it in
the HTML, with CSS styling similar to the old CSS-numbered chapters.
Books with a printStyle attribute like ccap-* have continuously
numbered chapters, all bottom-level entities are numbered #.#, and
higher level entities are just #.
Non-Ccaa books are numbered old-style, with all levels of numbers (e.g.
5 > 5.1 > 5.1.1)