Use iso code instead of array index for languages in advanced search (#1434)
On cnx.org, when we use advanced search to search for content in a
particular language, instead of doing /search?q=language:da, we're taken
to /search?q=language:6 and get no results.

When search languages are filtered, "languages" is an array and "@key" is
the array index.  This commit changes the code to use "id" instead,
which is the iso code.  (See src/scripts/collections/languages.coffee)