Use a staged build with the nginx container
This moves away from using nodejs with a manual nginx install. Instead we build
in the nodejs container and serve from the nginx container.

An entrypoint script is used to configure the container at runtime, which is when we have all the necessary information to configure both nginx and webview to point at the correct service domains.