A staged container build that is configurable at runtime (#1968)
* Use a staged build with the nginx container

This moves away from using nodejs with a manual nginx install. Instead we build
in the nodejs container and serve from the nginx container.

An entrypoint script is used to configure the container at runtime, which is when we have all the necessary information to configure both nginx and webview to point at the correct service domains.

* Fix to remove EOF newline

By removing this we can make the file look exactly liked the checked in version.

* Tell the user about the runtime environment

With this bipolar approach to serving the content for dev and prod, it's informative to have the runtime tell you what it's actually serving.

* Adjust nginx config generator for dev vs prod

This splits the dev configuration from that used by the 'prod' configuration.
This effectively allows the dev and prod versions to work in the browser.
And as far as I can tell, it allows the tests to run.

* Have docker ignore the build files

Docker will build it's own files from during container build-time. Also, the build files can be shared in a development sense by mounting the volume to the container at runtime.

* Wrap webview build script for env specific run

This wraps the webview scripts so we can conditionally run the build script.
It's not used in a dev environment, so no reason to run it.

* Fix to file permission changes statement

Don't change the file permissions when using as a 'dev' environment, because the dev environment usually means the developer has volume mounted their local code. Changing the permissions would therefore mean we'd change the files on the developers machine.

* Adjust docker-compose config based on Dockerfile changes