Add jenkins pipeline stages to deploy and integration test (#1974)
* Add jenkins pipeline stages to deploy and integration test

This will update the dev stack deployment of webview then proceed to run the integration tests over the newly deployed service.

* Label the deployment by the git commit hash

* Define the docker host to deploy to

* Use script argument instead of env var

Using the DOCKER_HOST variable is problematic, because it interacts with
all the docker executions, which isn't ideal. So we'll pass in the docker
host as an argument for determining the destination hostname.

* Fix report.xml generation to place it in a writable location

Pytest was having permission problems writing report.xml due to the container running as a user the host system doesn't know about. This adjustment moves the file into the jenkins job workspace after the tests have been run.