Issue 1944v2 (#1975)
* replace <br> and </p> tags with spaces

* add ellipsis and remove .more button when desc is short enought

* change elipsis (...) to (…)

* merge new changes from katAW.senna.syrena (#1967)

* remove border on focus
* show outline when moving with tab
* fixed book description on mobile
* issue 1905 show links in book descriptions (#1949)
* show links in description of books
* remove empty space
* issue 1757 read solution after open (#1961)
* read solution after open
* use role="alert" instead creating new element
* issue 1944 add more accessibility to toc (#1946)
* issue 1944 add more accessibility to toc
* aria-live for toc elements
* change .section-wrapper to button

* set role=alert dynamically to expanded elements

* add role=alert only for given module inside toc in sidebar

* Add jenkins pipeline stages to deploy and integration test (#1974)

* Add jenkins pipeline stages to deploy and integration test

This will update the dev stack deployment of webview then proceed to run the integration tests over the newly deployed service.

* Label the deployment by the git commit hash

* Define the docker host to deploy to

* Use script argument instead of env var

Using the DOCKER_HOST variable is problematic, because it interacts with
all the docker executions, which isn't ideal. So we'll pass in the docker
host as an argument for determining the destination hostname.

* Fix report.xml generation to place it in a writable location

Pytest was having permission problems writing report.xml due to the container running as a user the host system doesn't know about. This adjustment moves the file into the jenkins job workspace after the tests have been run.

* Fix and move the publish release stage (#1983)

Fix the "Publish Release" stage's condition by using Jenkins' built-in conditional functions.

Move this to the last stage in the pipeline to prevent premature release.

This also fixes the container tag name by dropping the `v` prefix.

* requested changes

* Update src/scripts/modules/media/nav/nav.coffee

* fix travis