Do a hard request for featured book clicks
We are setting the `data-bypass` attribute on the archor to allow the default
behavior to proceed. This passes the requested link to the browser rather
than having the javascript capture it for internal requesting of new content.

We make the browser do a request of these URLs so that it contacts
the webserver (nginx) to do redirects of REX books (aka all openstax books).