Allow specifying different hosts for archive depending on use
The FE_ARCHIVE environment variable is accepted for the docker container.
The idea is that FE_ARCHIVE will be used for frontend requests, whereas
ARCHIVE (with no prefix) will be used for nginx proxying inside the
container. The reason for this is that containers don't always recognize
the same web hosts as the docker hosts.

An example:
    FE_ARCHIVE=archive-local.cnx.org  # maybe in /etc/hosts
    FE_ARCHIVE_PORT=80
    ARCHIVE=archive  # recognized as host in the docker-compose network
    ARCHIVE_PORT=6543